Fix comment re: OscalFactory.
